+++ ReleaseNotes.md
@@ -1,3 +1,6 @@
+### New in 0.0.7 (Released 2015/07/29)
+* Rename PublishSettings to DeploySettings to prevent a confliction with "Microsoft.Web.Deployment"
+
 ### New in 0.0.6 (Released 2015/07/28)
 * Added namespace imports
 
@@ -13,8 +16,8 @@
 
 ### New in 0.0.3 (Released 2015/06/17)
 * Updated documentation
-* Added PublishSettings extensions methods
-* Add custom ports to PublishSettings
+* Added DeploySettings extensions methods
+* Add custom ports to DeploySettings
 
 ### New in 0.0.2 (Released 2015/06/17)
 * Added website test project to demostrate publish and deploy.